backend/requirements.txt
# FastAPI and ASGI
fastapi>=0.109.0
mangum>=0.17.0
uvicorn>=0.27.0

# AWS
boto3>=1.34.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# Utilities
python-multipart>=0.0.6
